Quaternions double->float (dunno how I missed that)

+++ src/main/java/org/getspout/api/math/MathHelper.java
@@ -139,7 +139,7 @@ public class MathHelper {
 	 * @param yaw
 	 * @return
 	 */
-	public static Vector3 getDirectionVector(double pitch, double yaw) {
+	public static Vector3 getDirectionVector(float pitch, float yaw) {
 		return Vector3.transform(Vector3.UNIT_X, Matrix.rotate((new Quaternion(pitch, Vector3.UNIT_Y)).multiply(new Quaternion(yaw, Vector3.UNIT_Z))));
 	}
