+++ src/main/java/org/getspout/api/math/Quaternion.java
@@ -6,7 +6,7 @@ package org.getspout.api.math;
  * 
  */
 public class Quaternion {
-	double x,y,z,w;
+	float x,y,z,w;
 	
 	/**
 	 * Represents no rotation
@@ -22,7 +22,7 @@ public class Quaternion {
 	 * @param z
 	 * @param w
 	 */
-	public Quaternion(double x, double y, double z, double w){
+	public Quaternion(float x, float y, float z, float w){
 		this.x = x;
 		this.y = y;
 		this.z = z;
@@ -33,8 +33,8 @@ public class Quaternion {
 	 * @param angle Angle, in Degrees, to rotate the axis about by
 	 * @param axis
 	 */
-	public Quaternion(double angle, Vector3 axis){
-		this(axis.getX() * MathHelper.sin(Math.toRadians(angle/2)), axis.getY() * MathHelper.sin(Math.toRadians(angle/2)), axis.getZ() * MathHelper.sin(Math.toRadians(angle/2)), MathHelper.cos(Math.toRadians(angle/2)));
+	public Quaternion(float angle, Vector3 axis){
+		this(axis.getX() * (float)MathHelper.sin(Math.toRadians(angle/2)), axis.getY() * (float)MathHelper.sin(Math.toRadians(angle/2)), axis.getZ() * (float)MathHelper.sin(Math.toRadians(angle/2)), (float)MathHelper.cos(Math.toRadians(angle/2)));
 	}
 	/**
 	 *  Copy Constructor
@@ -46,35 +46,35 @@ public class Quaternion {
 	 * Returns the X component of the quaternion
 	 * @return
 	 */
-	public double getX() {
+	public float getX() {
 		return x;
 	}
 	/**
 	 * Returns the Y component of the quaternion
 	 * @return
 	 */
-	public double getY() {
+	public float getY() {
 		return y;
 	}
 	/**
 	 * Returns the Z component of the quaternion
 	 * @return
 	 */
-	public double getZ() {
+	public float getZ() {
 		return z;
 	}
 	/**
 	 * Returns the W component of the quaternion
 	 * @return
 	 */
-	public double getW() {
+	public float getW() {
 		return w;
 	}
 	/**
 	 * Returns the length squared of the quaternion
 	 * @return
 	 */
-	public double lengthSquared(){
+	public float lengthSquared(){
 		return Quaternion.lengthSquared(this);
 	}
 	/**
@@ -82,7 +82,7 @@ public class Quaternion {
 	 * Note: This uses square root, so is slowish
 	 * @return
 	 */
-	public double length(){
+	public float length(){
 		return Quaternion.length(this);
 	}
 	/**
@@ -106,7 +106,7 @@ public class Quaternion {
 	 * @param axis
 	 * @return
 	 */
-	public Quaternion rotate(double angle, Vector3 axis){
+	public Quaternion rotate(float angle, Vector3 axis){
 		return Quaternion.rotate(this, angle, axis);
 	}
 	/**
@@ -128,7 +128,7 @@ public class Quaternion {
 	 * @param a
 	 * @return
 	 */
-	public static double lengthSquared(Quaternion a){
+	public static float lengthSquared(Quaternion a){
 		return (a.x*a.x + a.y*a.y + a.z*a.z + a.w*a.w);
 	}
 	/**
@@ -137,8 +137,8 @@ public class Quaternion {
 	 * @param a
 	 * @return
 	 */
-	public static double length(Quaternion a){
-		return MathHelper.sqrt(lengthSquared(a));
+	public static float length(Quaternion a){
+		return (float)MathHelper.sqrt(lengthSquared(a));
 	}
 	/**
 	 * Constructs and returns a new Quaternion that is the given Quaternion but length() == 1
@@ -146,7 +146,7 @@ public class Quaternion {
 	 * @return
 	 */
 	public static Quaternion normalize(Quaternion a){
-		double length = length(a);
+		float length = length(a);
 		return new Quaternion(a.x / length, a.y / length, a.z / length, a.w / length);
 	}
 	/**
@@ -156,10 +156,10 @@ public class Quaternion {
 	 * @return
 	 */
 	public static Quaternion multiply(Quaternion a, Quaternion b){
-		double x = a.w * b.x + a.x * b.w + a.y * b.z + a.z * b.y;
-		double y = a.w * b.y - a.x * b.z + a.y * b.w + a.z * b.x;
-		double z = a.w * b.z + a.x * b.y - a.y * b.x + a.z * b.w;
-		double w = a.w * b.w - a.x * b.x - a.y * b.y - a.z * b.z;
+		float x = a.w * b.x + a.x * b.w + a.y * b.z + a.z * b.y;
+		float y = a.w * b.y - a.x * b.z + a.y * b.w + a.z * b.x;
+		float z = a.w * b.z + a.x * b.y - a.y * b.x + a.z * b.w;
+		float w = a.w * b.w - a.x * b.x - a.y * b.y - a.z * b.z;
 		
 		return new Quaternion(x,y,z,w);
 	}
@@ -170,7 +170,7 @@ public class Quaternion {
 	 * @param axis
 	 * @return
 	 */
-	public static Quaternion rotate(Quaternion a, double angle, Vector3 axis){
+	public static Quaternion rotate(Quaternion a, float angle, Vector3 axis){
 		return multiply(new Quaternion(angle, axis), a);
 	}
 	/**
